Refactort reports as periodic tasks.
