+++ supervisord.conf
@@ -34,7 +34,7 @@ stderr_logfile=/dev/stderr
 stderr_logfile_maxbytes=0
 
 [program:reports]
-command=python3 create_reports.py --config %(ENV_CONFIG_FILE)s
+command=python3 run_periodic_task.py --config %(ENV_CONFIG_FILE)s
 stdout_logfile=/dev/stdout
 stdout_logfile_maxbytes=0
 stderr_logfile=/dev/stderr
